Unwrap errors for 403 responses as well
When we were refreshing the access tokens previously, then we would only interpret responses
with codes 200 or 400 as OAuth-specific ones (the latter was suggested in the spec).
However it looks like Auth0 uses HTTP 403 to indicate issues with the refresh token and
thus we need to try decoding OAuth errors from those as well. Not doing so would not allow
the client to stop refresh attempts when they make no sense anymore.